+++ qa/rpc-tests/p2p_txexpiry_dos.py
@@ -5,45 +5,14 @@
 
 import sys; assert sys.version_info < (3,), ur"This script does not run under Python 3. Please use Python 2.7.x."
 
-from test_framework.mininode import NodeConn, NodeConnCB, NetworkThread, \
-    CTransaction, msg_tx, mininode_lock, OVERWINTER_PROTO_VERSION
+from test_framework.mininode import NodeConn, NetworkThread, \
+    msg_tx, OVERWINTER_PROTO_VERSION
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import initialize_chain_clean, start_nodes, \
     p2p_port, assert_equal
+from tx_expiry_helper import TestNode, create_transaction
 
-import time, cStringIO
-from binascii import hexlify, unhexlify
-
-
-class TestNode(NodeConnCB):
-    def __init__(self):
-        NodeConnCB.__init__(self)
-        self.create_callback_map()
-        self.connection = None
-
-    def add_connection(self, conn):
-        self.connection = conn
-
-    # Spin until verack message is received from the node.
-    # We use this to signal that our test can begin. This
-    # is called from the testing thread, so it needs to acquire
-    # the global lock.
-    def wait_for_verack(self):
-        while True:
-            with mininode_lock:
-                if self.verack_received:
-                    return
-            time.sleep(0.05)
-
-    # Wrapper for the NodeConn's send_message function
-    def send_message(self, message):
-        self.connection.send_message(message)
-
-    def on_close(self, conn):
-        pass
-
-    def on_reject(self, conn, message):
-        conn.rejectMessage = message
+import time
 
 
 class TxExpiryDoSTest(BitcoinTestFramework):
@@ -55,24 +24,6 @@ class TxExpiryDoSTest(BitcoinTestFramework):
     def setup_network(self):
         self.nodes = start_nodes(1, self.options.tmpdir,
                                  extra_args=[['-nuparams=5ba81b19:10']])
-
-    def create_transaction(self, node, coinbase, to_address, amount, txModifier=None):
-        from_txid = node.getblock(coinbase)['tx'][0]
-        inputs = [{"txid": from_txid, "vout": 0}]
-        outputs = {to_address: amount}
-        rawtx = node.createrawtransaction(inputs, outputs)
-        tx = CTransaction()
-
-        if txModifier:
-            f = cStringIO.StringIO(unhexlify(rawtx))
-            tx.deserialize(f)
-            txModifier(tx)
-            rawtx = hexlify(tx.serialize())
-
-        signresult = node.signrawtransaction(rawtx)
-        f = cStringIO.StringIO(unhexlify(signresult['hex']))
-        tx.deserialize(f)
-        return tx
 
     def run_test(self):
         test_node = TestNode()
@@ -98,13 +49,11 @@ class TxExpiryDoSTest(BitcoinTestFramework):
         self.nodeaddress = self.nodes[0].getnewaddress()
 
         # Mininodes send transaction to zcashd node.
-        def setExpiryHeight(tx):
-            tx.nExpiryHeight = 101
-
-        spendtx = self.create_transaction(self.nodes[0],
-                                          self.coinbase_blocks[0],
-                                          self.nodeaddress, 1.0,
-                                          txModifier=setExpiryHeight)
+        spendtx = create_transaction(self.nodes[0],
+                                     self.coinbase_blocks[0],
+                                     self.nodeaddress,
+                                     1.0,
+                                     101)
         test_node.send_message(msg_tx(spendtx))
 
         time.sleep(3)
